# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DSS Backend API",
        "description": "KUSRC - DSS Backend API V0.0.3 (Update 18-09-2024 04:20AM)",
        "version": "0.0.3"
    },
    "servers": [
        {
            "url": "http://172.16.130.54:8000"
        }
    ],
    "paths": {
        "/api/getInventories": {
            "get": {
                "tags": [
                    "Inventories"
                ],
                "summary": "GET: Inventories Data with Pagination",
                "operationId": "740e5561e8eccb7f1c26a905459ade62",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/getInventories/{id}": {
            "get": {
                "tags": [
                    "Inventories"
                ],
                "summary": "GET: Inventories Data by ID",
                "operationId": "18f2af10e9724a33284b6804f7343222",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Inventory ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/getAllInventories": {
            "get": {
                "tags": [
                    "Inventories"
                ],
                "summary": "GET: All Inventories Data",
                "operationId": "e956a71e666d28bc0d863bc38dd6b370",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/getAllInventories/{id}": {
            "get": {
                "tags": [
                    "Inventories"
                ],
                "summary": "GET: All Inventories Data by ID",
                "operationId": "8caef8deee26b19cd7780ea2b380e773",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Inventory ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/getQrCodeInventories": {
            "get": {
                "tags": [
                    "Inventories"
                ],
                "summary": "GET: QR Code Inventories Data",
                "operationId": "6aa8f9260b2c4132e7a20ab473e55996",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/getCountInventories": {
            "get": {
                "tags": [
                    "Inventories"
                ],
                "summary": "GET: Count All Inventories Data",
                "operationId": "d7e903dc9c789a7585e26817f53108e2",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/putInventories/{id}": {
            "put": {
                "tags": [
                    "Inventories"
                ],
                "summary": "PUT: Update Inventory Data by ID",
                "operationId": "208c318d531d20254024e0bc74b38d38",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "check_fiscal_year": {
                                        "type": "integer",
                                        "example": 2567
                                    },
                                    "dispense_flg": {
                                        "type": "char",
                                        "example": "N"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
...
}